feat: disable-unrelated-to-gradle-hooks
